requirements.txt
flask
flask_cors
psycopg2
pyjwt
